Type synonym support
This changes the standard semantics to support type synonyms, which is one of
the most common feature requests.  Here are some example requests:

* https://github.com/dhall-lang/dhall-haskell/issues/10
* https://github.com/dhall-lang/dhall-haskell/issues/55
* https://github.com/dhall-lang/dhall-haskell/issues/92
* https://github.com/dhall-lang/dhall-haskell/issues/163
* https://github.com/dhall-lang/dhall-haskell/issues/176
* https://github.com/dhall-lang/dhall-haskell/issues/201
* https://github.com/dhall-lang/dhall-lang/issues/10

The change to the semantics is pretty simple: don't require that `let`
expressions type-check as an equivalent anonymous function.  Instead, just
perform substitution and type-check the substituted expression.  The semantics
are still sound and this behaves the way users expect.

For example, the following expression would be rejected before this change
and accepted after this change:

```haskell
let t = Integer in 1 : t
```