Fix semantics of `{Optional,List}/build`
Related to https://github.com/dhall-lang/dhall-haskell/issues/291

The old semantics produced a malformed expression if the user tried to use
`List/build` or `Optional/build` when the element type had a free variable
named `a`.  This fixes that to correctly shift free variables named `a` so
that they are not captured by the bound variable introduced in the reduction
rules for `Optional/build` or `List/build`

The old semantics also had useless fusion rules that would never fire because
an expression of the form `{Optional,List}/build A g` is not in normal form,
so this change also simplifies the semantics by removing those rules in order
to avoid confusion.