Fix and simplify typing judgments for record literals
The typing judgments for record literals were previously defined in terms of
the typing judgments for record types.  However, if you implement the
judgments literally this leads to an inefficient implementation since
the types of all fields are checked twice.  This change fixes that by
changing the judgments to only reference themselves when recursing over
the rest of the record.

This change also fixes a small bug in the specification for records that
store types because the specification did not require that the first
field in a record of types had type `Type`