Disallow `?` and `#` characters from paths
I discovered this issue in the grammar when updating the Haskell
implementation to match the standard import system.

The standardized import system changed the grammar of URLs so that the
path of a URL used the same grammar as a file path.  However, file paths
permitted `#` and `?` characters, leading to an ambiguity in how to
intepret those characters.

For example, in the following URL:

```
https://example.com/foo?bar=baz
```

... the grammar allowed the `?bar=baz` to be parsed as part of the
path instead of as a query parameter.

This change updates the grammar of paths (both for file paths and URLs)
to disallow the `?` and `#`.  This allows URLs and file paths to still
share the same grammar without leading to ambiguity when parsing query
parameters or fragments in a URL.