Add section about canonicalization of text
When looking into how to define the relating between single-line and
multiline string literals, I realised that text is defined only in terms
of its literals, which makes it awkward to define the relation between
the different text literals. I want to define them in terms of their
values, but there are no values in the standard, only literals. By
introducing a canonical encoding of a text value as a literal, we get a
way to define this relation.

As a side effect, the following two expressions are now equivalent:

    "\n"
    "\u000A"

And these two as well:

    "\uD834\uDD1E"
    "𝄞"

Previously this was not the case, because text literals were in normal
form.