Replace "pwd" job with "rev" job (#221)
See change to comment in source code for rationale. This avoids a copy, but is also more robust. We had a situtation at work where the previous approach didn't work because merging `master` in to a PR didn't actually change anything at all, but GitHub never received a status report.