Standardize how to encode/decode multi-`let` expressions
In https://github.com/dhall-lang/dhall-lang/pull/266 I forgot to specify how to
encode `let` expressions with multiple bindings.  This changes the encoding
to encode a `let` expression with multiple bindings as a single list with
three elements per binding.

This is not compatible with the previous encoding which would omit the type
annotation when absent instead of using `null`