Prevent Hurkens' paradox
Fixes https://github.com/dhall-lang/dhall-lang/issues/250

This uses a variation on CCω with two modifications:

* The lowest universe is still impredicative
* Dependent types are still forbidden

This suffices to prevent the paradox while still allowing support for
records of types of mixed kinds.

The main impact this has on other type-checking rules is that a record
of types is now a kind instead of type.  Or in other words, the type of
type of such a record is now `Sort` instead of `Kind`.  This arises
from the equivalent Boehm-Beraducci encoding of such a record.

The type-checking judgments for field access are not affected and the
equivalent Boehm-Berarducci encoding behaves the same as before.

This extends the test suite to check that records of mixed types are
supported and so are their equivalent encodings.