Add parser tests (#276)
Add a set of agnostic tests for the parser. Since the AST is not
standard, the closest thing we have is the binary encoding specification
(that is, the conversion to CBOR datastructures).

In this way we can verify that the parser is working correctly by
parsing expressions and then encoding them in their CBOR representation
(in plaintext JSON - for readability), and then checking that the
encoding is correct.