Disambiguate `#`/`?` operators from HTTP fragments/queries
Fixes https://github.com/dhall-lang/dhall-lang/issues/285

This adds required whitespace after the `#` and `?` operators so that they
cannot be confused with URL fragments or query parameters.

This also disambiguates alternative order for future reference so that
we have a well-defined behavior until we can iron out all of the ambiguous
corners of the grammar.