Rework summary
This is an attempt to improve Dhall's messaging by following some
of the advice in [Crossing the Chasm](https://en.wikipedia.org/wiki/Crossing_the_Chasm)

Specifically, the new summary answers the following two questions:

* Who is the intended audience?  (i.e. what the book calls "market positioning")
* How does Dhall improve upon the alternatives? (i.e. what the book calls "product positioning")

You can think of "market positioning" and "product positioning" in
terms of the cliche "it's like X but for Y" slogan (i.e. "it's like Uber
but for walking dogs").  The "X" encompasses the product positioning
by mentioning an analogous product that excels in the same way but
for a different market.  The "Y" encompasses the market positioning,
typically by mentioning the current market leader that you are trying
to displace so that people can instantly tell if they are the target audience.

In this case, the market we are addressing is people who are
maintaining large and unwieldy configuration files and YAML is the
market leader here although JSON is a close second.  Think:
Kubernetes, Ansible, Docker Compose.  Many people compare
Dhall to YAML and recommend Dhall as a YAML alternative.  Also,
pretty much every common use case for YAML is also an appropriate
use case for Dhall.

For the product positioning I chose Elm because I
think it is the closest analog to Dhall outside of the world of
configuration languages.  Both Elm and Dhall have a high
power-to-weight ratio and the comparison to Elm
concisely highlights how Dhall improves upon the alternative
configuration languages (types + functions + great error messages).