Disallow labels that match builtins or keywords
Fixes https://github.com/dhall-lang/dhall-haskell/issues/683

Before this change the grammar allowed variables to be named after
builtins or keywords, such as:

```haskell
let if = 1 in if
```

However, permitting this deteriorates error messages since an expression
like this:

```haskell
if a then b else
```

... becomes a type error ("Unbound variable: if") instead of a parse
error due to an incomplete `if` expression.

The solution is to disallow these sorts of variable names.  Given that
we don't have a good way to do that in the grammar, we just specify it
as a comment for now.