Standardize semantics for multi-line literals
Fixes #262

Supersedes #214

This upstreams the Haskell implementation's support for multi-line
strings into the standard and fixes several inconsistencies along
the way.

The main differences between the Haskell implementation and this
standardized version are:

* Newlines are mandatory after the opening `''` of a multi-line literal
* The trailing spaces before the closing `''` count towards the
  computation of leading indentation

The main difference between this and the first attempt to standardize
this in #214 is that normalization does not yet canonicalize escape sequences
as part of normalizing `Text` literals.  If we do want to support canonicalizing
escape sequences then that should be part of a separate pull request since not
even the Haskell implementation did that.  The scope of this pull request is to
bring the standard and Haskell implementation closer to parity.

This also fixes a few omissions and mistakes regarding string
interpolation in the `./standard/semantics.md` document that I ran
across.