Fix parsing test example
... as caught by @f-f

Fixes https://github.com/dhall-lang/dhall-lang/issues/343

According to the standard the expression:

```haskell
[1, 2, 3] : List Integer
```

... should be parsed as a list literal with a type annotation and encoded
that way.  The test example before this change incorporated a bug in the
Haskell code that encoded the type annotation as part of the list literal
alongside the list elements, which was not standards compliant.  This
change fixes that.