fix types in parser tests
The parser tests had a number of cases where they had Natural and
Integer syntax swapped round.  This presumably dates from before #138
switched the grammar of Integer and Natural.

Note that this difference doesn't actually matter for these parsing
tests, because an expression like `+1 : Natural` still parses correctly,
it just fails to typecheck.  But I found it confusing to have the syntax
so consistently backwards in the tests, so I thought it was worth
changing.