Simplify import resolution judgment
I realized that there was redundant information being threaded through
the import resolution judgment.  Specifically the `@ here` import always
matched the last import in the history of visited imports, so there is
no need to keep track of it separately from the history.

This change removes that redundancy by eliminating the `@ here` component
of the import resolution judgment and instead pattern matching on the
"head" of the non-empty set of visited imports to obtain the same import.