Expand character set for quoted labels
This expands quoted labels to permit all non-control ASCII characters except
backticks

One motivation for this is that I would like to eventually standardize support
for automatically converting nullary union alternatives to `Text`, but that
in turn requires expanding the set of valid alternative names.

A second motivation for this is to expand the set of field names that Dhall
can generate when targeting JSON/YAML configurations.