Update normalization tests
Fixes https://github.com/dhall-lang/dhall-lang/issues/413

This updates the normalization tests so that the `*B.dhall` files match the
output of the Haskell implementation.

In some cases this just changes the format of the file, but in other cases
this detects a case where the `*B.dhall` file was not in normal form.