Convert ABNF to prefix-whitespace (not suffix)
Hopefully fixes #395.

This implements an idle thought I had in
https://github.com/dhall-lang/dhall-lang/issues/395#issuecomment-471346496

> I have idly wondered if a PEG-compatible ABNF grammar might still be
> possible if we flipped it: if every expression consumed leading
> whitespace (not trailing), then higher-level rules like
> application-expression could impose extra whitespace restrictions in a
> PEG-compatible way.

This commit converts the ABNF rules from preferring trailing whitespace
to preferring leading whitespace.

This means that rules like application-expression will parse correctly
in a PEG.  To demonstrate the problem, consider an expression like
` f 3 `.  The first part, `f`, is a label and matches the `label` rule.
However a PEG is greedy and will consume all of the trailing whitespace.

Now, when you try to match application-expression, it will fail, because
there is no whitespace left for the required whitespace-chunk to match
on.  So the parse fails.

So far, PEG users have fixed this by removing the required whitespace
from the rule.  This does not introduce ambiguity because PEGs are
inherently unambiguous.  However, it means that PEG-based grammars will
parse expressions that CFG-based grammars won't, such as
`(Natural/even)2`.

By flipping the rules round to have leading whitespace instead of
trailing whitespace, this problem doesn't arise.

I have not tested these changes out in their entirety, but I have tested
out a subset of them in philandstuff/dhall-golang@1f6b6cd758 and it
passed all of the tests that had previously passed.