Fix parsing test for large expression
... as part of https://github.com/dhall-lang/dhall-lang/pull/438 I
mistakenly changed the large expression binary test result, even though it
did not need to be changed.  I don't remember exactly what confusion of events
led to me to computing the wrong golden test result while working on that
change.

This change reverts it back to where it was before.