add bool-lit grammar rule
This adds explicit rules for True and False to be parsed as bool
literals.  This is so that we don't parse them as an AST node that
represents an identifier the way we can parse all other builtins.

True and False are special because their CBOR representation isn't the
same as serialising the equivalent identifier.

tests/parser/success/builtinsA.dhall already checks that when we parse
True and serialise to CBOR, we do not serialise it the way we serialise
the equivalent identifier.

Fixes #483.