Seperate binary encoding of variables vs built-ins
Fixes #462

Built-ins always reference their built-in value, even when there is a
bound variable of the same name.  Bound variable with names matching
built-ins are allowed when quoted, but the quoting is not preserved in
the binary encoding.  This change ensures that the difference references
are encoded differently.

This will not change semantic hashes at all, since one alpha-normalizes
before hashing.