Simplify bare interpolations
Fixes https://github.com/dhall-lang/dhall-lang/issues/505

This upstreams a feature of the Haskell implementation to simplify
interpolated expressions if they are the sole contents of the `Text`
literal.

In other words, an expression of the form `"${e}"` simplifies further
to just `e`, even if `e` is a bound variable.