Remove referential sanity check
As I was implementing imports, I realised that the referential sanity
check is strictly more permissive than the CORS check.  This means we
don't need it, because anything which is not referentially sane will
also fail the CORS check.

This makes the import semantics simpler to describe, without changing
their meaning.

I also changed the "referentiallyInsane" test to try to use less of the
"referential sanity" language, since that terminology is no longer
defined in the semantics.