add missing base cases for union type type inference
The only way for a union type to have type `Kind` was using the rule:

     Γ ⊢ T :⇥ Kind   Γ ⊢ < ts… > :⇥ Kind
     ───────────────────────────────────  ; x ∉ < ts… >
     Γ ⊢ < x : T | ts… > : Kind

but this requires, as input, a union type of type `Kind`.  With no base
case to start the process, there was no way to construct a union type of
type `Kind`.  The same was true for union types of type `Sort`.

Also, make it clear that it's not a type error for a union alternative
to be a `Kind` (and so for the union type itself to have type `Sort`).