fix union constructor type inference judgments
The union constructor type inference judgments didn't infer types for
expressions of the form

    let MaybeNat = < Just Natural | Nothing > in MaybeNat.Nothing

because they required the union type to already be in normal form.  This
fixes that.

It also removes a somewhat-duplicate judgment for union constructors,
which would be able to type check `u.x` provided that a) `x` selected a
union alternative with a type (not an empty alternative), and b) `u` was
a union type with a minimum of two alternatives.  Once we fix the main
judgments, this extra judgment is entirely redundant.