Improve use of unions in tests (#540)
* delete alternativesAreTypesA.dhall

this is a byte-for-byte duplicate of unionOfTypesA.dhall in the
same directory

* Stop using empty records as union alternatives

Now that we have empty union alternatives, we don't need to use empty
records as singleton union alternative types in our tests.  Our tests
are also example code, and we should aim to keep them idiomatic as the
language evolves.

* add missing merge test case

We didn't have a unit test for type inference on merging an empty
alternative.

* tests: don't use union literals unless necessary

Union literals are deprecated in favour of union constructors.
Therefore, tests should prefer union constructors over union literals
except in cases where union literals cannot be avoided.  This is mainly
in cases where union literals themselves are being tested.

I also changed some filenames to make clear when union literals are
under test.