import as any-label
This does not change what programs are valid Dhall, but only moves
something like `./t.dhall as JSON` from a syntax error to a semantics
error.  This would allow users to add custom import types in implementations
that wish to support it in the same way that we currently allow adding
custom builtins.