Encode integrity check as multihash
Using a byte string with internal label as much more compact than
storing base16 as unicode.  Re-using the multihash spec so we don't have
to invent our own but using such a small subset of it that implementors
do not need to be familiar with multihash at all to implement this.

This change does not affect semantic hashes, since semantic hashes are
computed on fully resolved expressions.

Closes #548