Promote some headings in README.md (#551)
I was looking at the README and it seemed strange that some of the most important information is nested under the "Overview" header. This commit promotes the "Interpreter", "Language Bindings", and "Compilers" sections. While these are all referenced from the "Overview" section, I don't think that means they should be nested within it.