More fixes to unit tests (#553)
* require `Natural/fold 1` to call its function

I implemented `Natural/fold` enough to pass NaturalFoldZeroA.dhall and
was surprised to see NaturalFoldOneA.dhall also passed.  This changes
NaturalFoldOneA.dhall so that it won't pass if `Natural/fold` is
implemented as

    Natural/fold _ _ _ x = x

which was my minimal implementation intended to pass
NaturalFoldZeroA.dhall.

* remove duplicate tests

these tests duplicate tests in the /unit/ directory.

* note that `Text/show` skips interpolated text

It wasn't particularly explict that `Text/show` only operates on text
literals containing no interpolations.  This makes it clearer.

Also, add an explicit test for this case.

* tests for List/{head,last}: test longer lists

If we only test singleton lists, we don't test that we're actually
fetching the correct element.

* fix typo in test