Fix tests multi-line strings with leading tabs
I believe that I ran across a bug in parsing test for `largeExpressionA.dhall`.
Specifically, that test was encoding the final multi-line string literal
to including the leading tabs, even though the standard specifies that
they should be stripped as part of dedenting the strings.

Specifically, the CBOR encoding said to encode the multi-line string as:

```
[18, "\n\t\t\t\t\t\t\t\t\t\t\t\t  "]
```

... when it should have been:

```
[18, "\n"]
```

However, I actually didn't expect there to be leading tabs in this
expression in the first place, so rather than just update the encoding I
instead reformatted the expression and created a separate test focused
on just testing support for dedenting leading tabs correctly.