improve alternativeHashMismatchA.dhall
This test imports its own expected output, with a hash mismatch.  This
isn't a great test, because an implementation which parses and then
completely ignores integrity checks will pass this test: it will resolve
the import, without error, and return the correct value.

This commit changes the test to import a different value.  The only way
to pass this test is now to correctly reject the first import due to
hash mismatch, and then use the import alternative operator to pick the
correct value instead.