add a bunch of binary decode tests
I'm just getting around to implement import caching, for which I need to
implement binary decoding.  As a result, this is an opportunity for me
to add a bunch of tests to the `binary-decode` suite.

This adds unit tests for all binary decode judgments except for:

- record projection (either `x.{a,b}` or `x.(e)`) -- I haven't yet
  implemented them in dhall-golang so I can't check my working
- union literals (again, I haven't implemented them, but they'll be
  removed soon anyway)

Broadly, the tests assume that the first things you implement binary
decoding for are:

 - builtins
 - underscore variables
 - named variables
 - Natural literals

Once you have these, you should probably be able to implement the rest
of the tests in any order.