Add standard representation for weakly-typed JSON values
This adds a new `JSON` type for representing unstructured JSON data
in Dhall.

The goal is to provide a way for users to have a way to pass through
unstructured data when integrating tools that have schema-free subsets
of their APIs.  An example of this is "pass-through" API fields which pass
through arbitrary JSON data without interpreting that data in any way.

For people familiar with Haskell, this newly-added JSON type is the Dhall
analog of `Data.Aeson.Value`.

This will be paired with a matching change to the `dhall-json` package
to recognize this type and support converting it back and forth to
`Data.Aeson.Value`s.