Use RFC3986 section 5 URL resolution algorithm
This is part 1 of the proposal in #581.  This changes the chaining
operation so that a parent URL chained to a child local import is now
defined in terms of the RFC3986 section 5 URL resolution algorithm.

I have erred on the side of being too verbose rather than not verbose
enough.

As part of this, I have started to use the notation `URL` (and `URL₀`,
`URL₁`, etc) to stand in for any generic URL import.  I think as things
progress on #581, URLs can be treated more opaquely and the current
`https://authority₀ directory₀ file₀` notation can be removed.

As part of this, I have also changed the internal representation of a
URL import to be a literal URL, rather than the parsed and decoded
components of a URL.  This was necessary, because without this change we
would do percent-encoding in two places: when converting a local import
to a relative reference, and when fetching a URL.  This meant that we
could end up double-percent-encoding a URL.  For example, if we chain
`https://example.com/foo` on to the local path `./" "`, we would first
convert the local path to `./%20`, resolve the relative reference to
`https://example.com/%20`, and then when fetching the URL,
percent-encode it again to `https://example.com/%2520`.  This is
incorrect behaviour.

After this change, we only perform percent-encoding either a) when
parsing a URL containing quoted path components; or b) when converting a
local import to a relative reference.  The same data cannot go through
both processes, so there is now no risk of double-percent-encoding.
Further, we never perform percent-decoding any more.

In particular:

 - the binary representation of a URL is now the CBOR-specified encoding
   of a URL (see RFC7049 section 2.4.4.3)

 - quoted path components in URLs should be percent-encoded at parse
   time and converted to a valid URL

This commit is a breaking change; however, its breakages are somewhat
esoteric.  The binary format has changed, but it won't affect cached
expressions or semantic hashes, because both are computed on
fully-resolved expressions, which don't contain imports.